IncindentHandler : listens (asks every 5min) the api for new incidents and creates and creates the incidents entities, TeamGenerator will ask for incidents and assign compatible teams

+++ Java/EmergencyManager/src/entities/Incident.java
@@ -16,7 +16,7 @@ public class Incident {
         this.id = id;
     }
 
-    public Incident(int id, float latitude, float longitude, int cityId, int intensity, Date updatedAt) {
+    public Incident(int id, float latitude, float longitude, int cityId, float intensity, Date updatedAt) {
         this.id = id;
         this.latitude = latitude;
         this.longitude = longitude;
